refactor: split mount.ts into component subfolder
Extract component lifecycle logic from mount.ts into render/component/:
- lifecycle.ts: shared helpers (processHookDescriptors, commitOrDefer, runComponentRender)
- initial.ts: initialComponentRender
- rerender.ts: executeComponentRerender, resumeInstance, rerenderInstance
- index.ts: createComponentInstance, mountComponent

Also: remove _ prefix from ComponentInstance.resume/executeRerender,
add mounted boolean, export CtxMap from driver.

Closes #157
