+++ src/render/component/index.ts
@@ -1,0 +1,69 @@
+/**
+ * component/index.ts — Component instance creation and mounting.
+ *
+ * Public entry points:
+ * - `createComponentInstance` — create a ComponentInstance with bound lifecycle closures.
+ * - `mountComponent` — create the instance and run its initial render.
+ */
+
+import type { Component, InternalProps } from "../../jsx";
+import { type CtxMap, drive, getContextMap, type RenderGenerator } from "../driver";
+import type { ComponentInstance } from "../types";
+import { initialComponentRender } from "./initial";
+import { effectiveCtxMap } from "./lifecycle";
+import { executeComponentRerender, rerenderInstance, resumeInstance } from "./rerender";
+
+/**
+ * Create a `ComponentInstance` with bound lifecycle closures.
+ * Pure synchronous work — no yields.
+ */
+function createComponentInstance(
+  component: Component,
+  props: InternalProps,
+  ctxMap: CtxMap,
+): ComponentInstance {
+  const instance: ComponentInstance = {
+    component,
+    props,
+    endMarker: document.createComment(""),
+    capturedCtx: ctxMap,
+    slots: [],
+    hookStates: [],
+    cleanupFns: [],
+    pendingEffects: [],
+    localPatchRefCount: 0,
+    mounted: false,
+    isRendering: false,
+    pendingRerender: false,
+    renderResolvers: [],
+    resumeHookIndex: 0,
+    consumedContexts: new Set(),
+    providedContexts: new Set(),
+    resume: () => {
+      drive(effectiveCtxMap(instance), resumeInstance(instance));
+    },
+    executeRerender: () => {
+      drive(effectiveCtxMap(instance), executeComponentRerender(instance));
+      return Promise.resolve();
+    },
+    rerender: () => rerenderInstance(instance),
+  };
+
+  return instance;
+}
+
+/**
+ * Mount a component: create the instance and run the initial render.
+ *
+ * Returns a `DocumentFragment` containing the output nodes and the
+ * instance for slot tracking.
+ */
+export function* mountComponent(
+  component: Component,
+  props: InternalProps,
+): RenderGenerator<{ fragment: DocumentFragment; componentInstance: ComponentInstance }> {
+  const ctxMap = yield* getContextMap();
+  const instance = createComponentInstance(component, props, ctxMap);
+  const fragment = yield* initialComponentRender(instance);
+  return { fragment, componentInstance: instance };
+}
